% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_internal_helpers.R
\name{add_export_menu}
\alias{add_export_menu}
\title{Add Export Menu to Highcharter Plots}
\usage{
add_export_menu(hc)
}
\arguments{
\item{hc}{A highcharter object.}
}
\value{
A highcharter object with export menu enabled.
}
\description{
Helper function to add consistent export/download options to highcharter plots.
}
\keyword{internal}
