% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_uniqueness.R
\name{consistency_panel_ui}
\alias{consistency_panel_ui}
\title{Consistency Panel UI}
\usage{
consistency_panel_ui(id, .data)
}
\arguments{
\item{id}{Character string. The module namespace ID.}

\item{.data}{Data frame. The data to be used in the consistency panel.}
}
\value{
A Shiny UI object representing the consistency panel.
}
\description{
Consistency Panel UI
}
