% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_coverage.R
\name{coverage_server}
\alias{coverage_server}
\title{Coverage Server Module}
\usage{
coverage_server(id, est_data, personnel_data, contract_data, cache)
}
\arguments{
\item{id}{Character string. The module namespace ID.}

\item{est_data}{Data frame. The establishment-level data.}

\item{personnel_data}{Data frame. The personnel-level data.}

\item{contract_data}{Data frame. The contract-level data.}

\item{cache}{A list of dataframes. It caches pre-computed coverage results. Passed on to \code{coverage_panel_server()}.}
}
\value{
None. Called for side effects (renders Shiny outputs).
}
\description{
Server logic for the coverage section, processing and visualizing missing data patterns.
}
\keyword{internal}
