% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govhrapp-package.R
\docType{package}
\name{govhrapp-package}
\alias{govhrapp}
\alias{govhrapp-package}
\title{govhrapp: A Shiny application suite to inform human resource management}
\description{
This package is developed by the World Bank. It enables users to quality check, analyze and forecast HR data to inform human resource management.
}
\author{
\strong{Maintainer}: Galileu Kim \email{galileukim@worldbank.org}

Authors:
\itemize{
  \item Galileu Kim \email{galileukim@worldbank.org}
  \item Ifeanyi Edochie \email{iedochie@worldbank.org}
}

}
\keyword{internal}
