% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_plots.R
\name{plot_density}
\alias{plot_density}
\title{Plot Density as Percentage Share}
\usage{
plot_density(.data, group_col, measure_col)
}
\arguments{
\item{.data}{A data frame.}

\item{group_col}{The column name to group by.}

\item{measure_col}{The column name of the numeric variable to plot.}
}
\value{
A plotly object.
}
\description{
Plot Density as Percentage Share
}
