% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_plots.R
\name{plot_segment}
\alias{plot_segment}
\title{Plot Value Ranges by Group}
\usage{
plot_segment(.data, measure_col, group_col)
}
\arguments{
\item{.data}{Data frame containing the variables to plot.}

\item{measure_col}{Character. Numeric column to plot on the x-axis.}

\item{group_col}{Character. Column to group by, plotted on the y-axis.}
}
\value{
A ggplot2 object.
}
\description{
Draws a grey range segment from each group's minimum to its maximum, overlaid
with jittered observations. Groups are ordered by mean value, highest at the
top.
}
\examples{
plot_segment(mtcars, measure_col = "mpg", group_col = "cyl")

}
