% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_plots.R
\name{plot_transfer_network}
\alias{plot_transfer_network}
\title{Plot Transfer Network}
\usage{
plot_transfer_network(.data)
}
\arguments{
\item{.data}{A data frame containing the transfer data with columns \code{from}, \code{to}, and \code{weight}.}
}
\value{
A ggplot2 object representing a transfer network.
}
\description{
Plot Transfer Network
}
