% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_plots.R
\name{plotly_transfer_network}
\alias{plotly_transfer_network}
\title{Plot Transfer Network (plotly implementation)}
\usage{
plotly_transfer_network(.data)
}
\arguments{
\item{.data}{A data frame containing the transfer data with columns \code{from} and \code{to}.}
}
\value{
A plotly object representing a transfer network arc diagram.
}
\description{
Plot Transfer Network (plotly implementation)
}
