% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_app.R
\name{run_qcheckapp}
\alias{run_qcheckapp}
\title{Run the Quality Assessment Shiny Application}
\usage{
run_qcheckapp(qc_obj, ...)
}
\arguments{
\item{qc_obj}{A quality control object created by \code{govhr::compute_qualitycontrol()}.}

\item{...}{Additional arguments passed to \code{\link[shiny]{shinyApp}}.}
}
\value{
A Shiny app object.
}
\description{
Launches an interactive Shiny dashboard for govhr data visualization and analysis.
}
\examples{
\dontrun{
qc_obj <- govhr::compute_qualitycontrol(
  contract_dt  = govhr::bra_hrmis_contract,
  personnel_dt = govhr::bra_hrmis_personnel,
  est_dt       = govhr::bra_hrmis_est
)
run_qcheckapp(qc_obj)
}

}
