% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill.R
\name{wagebill_app}
\alias{wagebill_app}
\title{Standalone Wage Bill Application}
\usage{
wagebill_app()
}
\value{
A Shiny app object.
}
\description{
Launches a standalone Shiny app for testing the wage bill module.
}
\examples{
\dontrun{
wagebill_app()
}
}
\keyword{internal}
