% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_equity.R
\name{wagebill_equity_ui}
\alias{wagebill_equity_ui}
\title{Wage Bill Equity UI}
\usage{
wagebill_equity_ui(id, .data)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{.data}{Data frame containing wage bill data.}
}
\value{
A Shiny UI definition.
}
\description{
Sidebar controls and plots for wage distribution, deciles and compression.
}
