% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_movement.R
\name{wagebill_movement_server}
\alias{wagebill_movement_server}
\title{Wage Bill Movement Server}
\usage{
wagebill_movement_server(id, .data, cache)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{.data}{Data frame containing wage bill data.}

\item{cache}{List of pre-computed summaries from \code{\link[=build_analytics_cache]{build_analytics_cache()}}.}
}
\value{
A Shiny module server function.
}
\description{
Wage Bill Movement Server
}
\keyword{internal}
