% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_movement.R
\name{wagebill_movement_ui}
\alias{wagebill_movement_ui}
\title{Wage Bill Movement UI}
\usage{
wagebill_movement_ui(id, .data)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{.data}{Data frame containing wage bill data.}
}
\value{
A Shiny UI definition.
}
\description{
Sidebar controls and plots for the cost of recruitment and separation.
}
