% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_overview.R
\name{wagebill_overview_ui}
\alias{wagebill_overview_ui}
\title{Wage Bill Overview UI}
\usage{
wagebill_overview_ui(id, .data)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{.data}{Data frame containing wage bill data.}
}
\value{
A Shiny UI definition.
}
\description{
Sidebar controls, a wage bill time trend, and totals and growth rates by
group.
}
