% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_retirement.R
\name{wagebill_retirement_server}
\alias{wagebill_retirement_server}
\title{Wage Bill Retirement Server}
\usage{
wagebill_retirement_server(id, .data, cache)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{.data}{Data frame containing wage bill data.}

\item{cache}{List of pre-computed summaries from \code{\link[=build_analytics_cache]{build_analytics_cache()}}.}
}
\value{
A Shiny module server function.
}
\description{
Wage Bill Retirement Server
}
\keyword{internal}
