% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_retirement.R
\name{wagebill_retirement_ui}
\alias{wagebill_retirement_ui}
\title{Wage Bill Retirement UI}
\usage{
wagebill_retirement_ui(id, .data)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{.data}{Data frame containing wage bill data.}
}
\value{
A Shiny UI definition.
}
\description{
Sidebar controls and plots for realised and projected retirement costs.
}
