% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_beta.R
\name{wagebill_server_beta}
\alias{wagebill_server_beta}
\title{Wage Bill Server Module}
\usage{
wagebill_server_beta(id, wagebill_data)
}
\arguments{
\item{id}{Module id.}

\item{wagebill_data}{Data frame with wage bill data.}
}
\description{
Server logic for wage bill analytics server
}
