% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_beta.R
\name{wagebill_ui_beta}
\alias{wagebill_ui_beta}
\title{Wage Bill UI Module}
\usage{
wagebill_ui_beta(id, wagebill_data)
}
\arguments{
\item{id}{Module id.}

\item{wagebill_data}{Data frame with wage bill data.}
}
\description{
UI for wage bill analytics, including overview, controls, and plots.
}
