% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_movement.R
\name{workforce_movement_server}
\alias{workforce_movement_server}
\title{Workforce Movement Server}
\usage{
workforce_movement_server(id, .data, cache)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{.data}{Data frame containing personnel data.}

\item{cache}{List of pre-computed summaries from \code{\link[=build_analytics_cache]{build_analytics_cache()}}.}
}
\value{
A Shiny module server function.
}
\description{
Workforce Movement Server
}
\keyword{internal}
