% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_overview.R
\name{workforce_overview_server}
\alias{workforce_overview_server}
\title{Workforce Overview Server}
\usage{
workforce_overview_server(id, .data, cache)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{.data}{Data frame containing workforce data.}

\item{cache}{List of pre-computed summaries from \code{\link[=build_analytics_cache]{build_analytics_cache()}}.}
}
\value{
A Shiny module server function.
}
\description{
Workforce Overview Server
}
\keyword{internal}
