% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_retirement.R
\name{workforce_retirement_ui}
\alias{workforce_retirement_ui}
\title{Workforce Retirement UI}
\usage{
workforce_retirement_ui(id, .data)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{.data}{Data frame containing personnel data.}
}
\value{
A Shiny UI definition.
}
\description{
Sidebar controls and plots for realised and projected retirements.
}
