% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_transfer.R
\name{workforce_transfer_server}
\alias{workforce_transfer_server}
\title{Workforce Transfer Server Module}
\usage{
workforce_transfer_server(id, .data, cache)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}

\item{cache}{A list containing pre-computed data for caching.}
}
\value{
A Shiny server module for workforce transfer analytics.
}
\description{
Workforce Transfer Server Module
}
