% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_transfer.R
\name{workforce_transfer_ui}
\alias{workforce_transfer_ui}
\title{Workforce Transfer UI Module}
\usage{
workforce_transfer_ui(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}
}
\value{
A Shiny module UI function for workforce transfer analytics.
}
\description{
Workforce Transfer UI Module
}
