% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_transition.R
\name{workforce_transition_app}
\alias{workforce_transition_app}
\title{Launch the Workforce Transition Module Standalone}
\usage{
workforce_transition_app(.data, cache)
}
\arguments{
\item{.data}{Data frame containing wage bill data.}

\item{cache}{List of pre-computed summaries from \code{\link[=build_analytics_cache]{build_analytics_cache()}}.}
}
\value{
A Shiny app object.
}
\description{
Launch the Workforce Transition Module Standalone
}
\keyword{internal}
