% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_transition.R
\name{workforce_transition_server}
\alias{workforce_transition_server}
\title{Workforce Transition Server}
\usage{
workforce_transition_server(id, .data, cache)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{.data}{Data frame containing wage bill data.}

\item{cache}{List of pre-computed summaries from \code{\link[=build_analytics_cache]{build_analytics_cache()}}.}
}
\value{
A Shiny module server function.
}
\description{
Workforce Transition Server
}
