% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_transition.R
\name{workforce_transition_ui}
\alias{workforce_transition_ui}
\title{Workforce Transition UI}
\usage{
workforce_transition_ui(id, .data)
}
\arguments{
\item{id}{Character. Module namespace ID.}

\item{.data}{Data frame containing wage bill data.}
}
\value{
A Shiny UI definition.
}
\description{
Sidebar controls, a transition trend plot and an interactive transition
network.
}
