% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_beta.R
\name{workforce_ui_beta}
\alias{workforce_ui_beta}
\title{Workforce Analytics Module (Beta)}
\usage{
workforce_ui_beta(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing personnel data.}
}
\value{
A Shiny app object for workforce analytics.
}
\description{
Workforce Analytics Module (Beta)
}
