% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{build_analytics_cache}
\alias{build_analytics_cache}
\title{Build cache for analytics app}
\usage{
build_analytics_cache(workforce_data, wagebill_data)
}
\arguments{
\item{workforce_data}{A data frame containing workforce data.}

\item{wagebill_data}{A data frame containing wage bill data.}
}
\value{
A list containing cached data frames for improved performance.
}
\description{
Build cache for analytics app
}
