% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_helpers.R
\name{build_wagebill_group_choices}
\alias{build_wagebill_group_choices}
\title{Build grouping variable choices for wagebill UI}
\usage{
build_wagebill_group_choices(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A named list of choices available in the data.
}
\description{
Constructs a named list of grouping choices for use in a Shiny selectInput,
grouped by dictionary module. Always includes an "All" option mapped to
\code{ref_date}.
}
\examples{
example_data <- data.frame(
"ref_date" = as.Date(c("2020-01-01", "2020-02-01")),
"personnel_id" = c(1, 2),
"gross_salary_lcu" = c(1000, 1500)
)

choices <- build_wagebill_group_choices(example_data)

}
