% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{build_wagebill_group_choices}
\alias{build_wagebill_group_choices}
\title{Build grouping variable choices for wagebill UI}
\usage{
build_wagebill_group_choices(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A named list of choices available in the data.
}
\description{
Constructs a named list of grouping choices for use in a Shiny selectInput,
grouped by dictionary module. Always includes an "All" option mapped to
\code{ref_date}.
}
\examples{
choices <- build_wagebill_group_choices(bra_hrmis_contract)

}
