% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govhr_port.R
\name{compute_compression_ratio}
\alias{compute_compression_ratio}
\title{Function to compute the compression ratio}
\usage{
compute_compression_ratio(
  .data,
  group_cols = NULL,
  percentiles = c(0.9, 0.5, 0.1),
  measure_col,
  latest_measure = FALSE
)
}
\arguments{
\item{.data}{A data frame.}

\item{group_cols}{A character vector of column names to group the data by.}

\item{percentiles}{A numeric vector of length 3 indicating the upper, middle, and lower percentiles to compute (default is c(0.9, 0.5, 0.1)).}

\item{measure_col}{The name of the column for which the compression ratio will be computed.}

\item{latest_measure}{A logical value indicating whether to return only the measures for the latest reference date.}
}
\value{
A data frame containing the 90th, 50th, and 10th percentiles for the specified measure column within the specified groups and reference dates.
}
\description{
Function to compute the compression ratio
}
