% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_helpers.R
\name{compute_global_coverage}
\alias{compute_global_coverage}
\title{a function to compute the proportion of missing values in a data frame}
\usage{
compute_global_coverage(data, digits = 2)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A numeric value representing the proportion of missing values in the data frame.
}
\description{
a function to compute the proportion of missing values in a data frame
}
