% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govhr_port.R
\name{compute_movement_cost}
\alias{compute_movement_cost}
\title{Function to compute the total cost associated with personnel movement events.}
\usage{
compute_movement_cost(
  .data,
  id_col = "personnel_id",
  event_type,
  start_date = NULL,
  end_date = NULL,
  status_col = "employment_status",
  freq = NULL,
  measure_col,
  group_cols = NULL,
  latest_measure = FALSE
)
}
\arguments{
\item{.data}{A data frame containing the data to be processed.}

\item{id_col}{The name of the column representing personnel IDs (default is "personnel_id").}

\item{event_type}{A character vector indicating which movement event(s) to include (e.g., "hire", "fire", "retirement"). Multiple types can be supplied to compute costs for each type.}

\item{start_date}{The start date for the classification period. Defaults to the minimum reference date found in \code{.data}.}

\item{end_date}{The end date for the classification period. Defaults to the maximum reference date found in \code{.data}.}

\item{status_col}{The name of the column representing employment status (default is "employment_status").}

\item{freq}{The frequency of the reference dates. Defaults to a guess based on \code{.data}.}

\item{measure_col}{The name of the column containing the cost/measure to sum.}

\item{group_cols}{A character vector of column names to group the data by.}

\item{latest_measure}{A logical value indicating whether to return only the measures for the latest reference date.}
}
\value{
A data frame containing the movement cost for each requested event type within the specified groups and reference dates.
}
\description{
Function to compute the total cost associated with personnel movement events.
}
