% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govhr_port.R
\name{compute_percentile}
\alias{compute_percentile}
\title{Function to compute the cumulative distribution function}
\usage{
compute_percentile(
  .data,
  group_col = NULL,
  measure_col,
  binwidth = 1,
  latest_measure = FALSE
)
}
\arguments{
\item{.data}{A data frame.}

\item{group_col}{A character vector of column names to group the data by.}

\item{measure_col}{The name of the column for which the percentile values will be computed.}

\item{binwidth}{The width of the bins for grouping the measure values (default is 1).}

\item{latest_measure}{A logical value indicating whether to return only the measures for the latest reference date.}
}
\value{
A data frame with the cumulative distribution function.
}
\description{
Function to compute the cumulative distribution function
}
