% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_helpers.R
\name{compute_value_consistency}
\alias{compute_value_consistency}
\title{Compute the proportion of consistent values in a data frame.}
\usage{
compute_value_consistency(
  data,
  id_col,
  value_col,
  group_cols = NULL,
  digits = 2
)
}
\arguments{
\item{data}{A data frame.}

\item{id_col}{A character string specifying the name of the column that uniquely identifies records.}

\item{value_col}{A character string specifying the name of the column whose values are to be checked for consistency.}

\item{group_cols}{A character vector specifying the names of the columns to group by. Default is no grouping.}

\item{digits}{An integer specifying the number of decimal places to round the result to. Default is 2.}
}
\value{
A data frame with the proportion of consistent values in the data frame, optionally by group.
}
\description{
Compute the proportion of consistent values in a data frame.
}
