% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_consistency.R
\name{consistency_panel_server}
\alias{consistency_panel_server}
\title{Consistency Panel Server}
\usage{
consistency_panel_server(id, .data, cache)
}
\arguments{
\item{id}{Character string. Sub-module ID.}

\item{.data}{A dataframe. Input dataset for the sub-module.}

\item{cache}{A list of dataframes. Precomputed consistency data for the default grouping (ref_date).}
}
\value{
A set of Shiny outputs for the consistency panel.
}
\description{
Consistency Panel Server
}
