% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_uniqueness.R
\name{consistency_ui}
\alias{consistency_ui}
\title{Consistency UI Module}
\usage{
consistency_ui(id, est_data, personnel_data, contract_data)
}
\arguments{
\item{id}{Character string. The module namespace ID.}

\item{est_data}{Data frame. The establishment-level data.}

\item{personnel_data}{Data frame. The personnel-level data.}

\item{contract_data}{Data frame. The contract-level data.}
}
\value{
A Shiny UI object containing coverage analysis cards.
}
\description{
UI components for the consistency section of the quality control dashboard.
}
\keyword{internal}
