% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_helpers.R
\name{coverage_panel_server}
\alias{coverage_panel_server}
\title{Coverage Panel Server}
\usage{
coverage_panel_server(id, .data)
}
\arguments{
\item{id}{Character string. Sub-module ID matching the one used in
\code{\link[=coverage_panel_ui]{coverage_panel_ui()}}.}

\item{.data}{Data frame for the panel (e.g., establishment, personnel, or contract).}
}
\value{
A set of Shiny outputs for the coverage panel.
}
\description{
Server logic for individual coverage panels.
}
