% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_helpers.R
\name{coverage_panel_ui}
\alias{coverage_panel_ui}
\title{Coverage Panel UI}
\usage{
coverage_panel_ui(id, .data)
}
\arguments{
\item{id}{Character string. The module namespace ID.}

\item{.data}{Data frame. The data to be used in the coverage panel.}
}
\value{
A Shiny UI object representing the coverage panel.
}
\description{
Coverage Panel UI
}
