% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{date_ui}
\alias{date_ui}
\title{UI for the date range input.}
\usage{
date_ui(id, .data)
}
\arguments{
\item{id}{Character string. The module namespace ID.}

\item{.data}{Data frame. The data to be used for determining the date range.}
}
\value{
A Shiny UI element for selecting a date range.
}
\description{
UI for the date range input.
}
