% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{default_ui_controls}
\alias{default_ui_controls}
\title{Generate UI controls to filter data.}
\usage{
default_ui_controls(.data, id)
}
\arguments{
\item{.data}{Data frame. The data to be filtered.}

\item{id}{Character string. The module namespace ID.}
}
\value{
A list of Shiny UI elements for filtering the data.
}
\description{
Generate UI controls to filter data.
}
