% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_helpers.R
\name{filter_data}
\alias{filter_data}
\title{Filter data based on group, subgroup, and date range inputs}
\usage{
filter_data(data, group_filter, subgroup_filter, date_range)
}
\arguments{
\item{data}{A data frame containing the data to be filtered.}

\item{group_filter}{A character string specifying the group variable to filter by.}

\item{subgroup_filter}{A character vector specifying the subgroups to filter by.}

\item{date_range}{A vector of two dates specifying the date range to filter by.}
}
\value{
A filtered data frame based on the specified inputs.
}
\description{
Filter data based on group, subgroup, and date range inputs
}
