% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{group_filter_ui}
\alias{group_filter_ui}
\title{UI for the group filter input.}
\usage{
group_filter_ui(id, .data, selected = "ref_date", group_choices = NULL)
}
\arguments{
\item{id}{Character string. The module namespace ID.}

\item{.data}{Data frame. The data to be used for determining the available grouping choices.}

\item{selected}{Character string. The default selected group. Defaults to "ref_date".}

\item{group_choices}{Optional named vector of grouping choices. If NULL, the function will identify available grouping choices from the data.}
}
\value{
A Shiny UI element for selecting a group filter.
}
\description{
UI for the group filter input.
}
