% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_helpers.R
\name{identify_group_choices}
\alias{identify_group_choices}
\title{Identify available grouping choices based on the columns present in the data.}
\usage{
identify_group_choices(.data)
}
\arguments{
\item{.data}{A data frame.}
}
\value{
A named list of grouping choices, where each element corresponds to a module and contains a named vector of variable IDs and their corresponding variable names.
}
\description{
Identify available grouping choices based on the columns present in the data.
}
