% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_utils.R
\name{identify_wagebill_choices}
\alias{identify_wagebill_choices}
\title{Function to identify available wagebill measure choices from the data and dictionary}
\usage{
identify_wagebill_choices(.data)
}
\arguments{
\item{.data}{A data frame containing wagebill data.}
}
\value{
A named list of wagebill measure choices, where each element corresponds to a module and contains a named vector of variable IDs and their corresponding variable names.
}
\description{
Function to identify available wagebill measure choices from the data and dictionary
}
