% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_overview.R
\name{overview_server}
\alias{overview_server}
\title{Overview Server Module}
\usage{
overview_server(id, meso_table)
}
\arguments{
\item{id}{Module id.}

\item{meso_table}{Data frame with meso-level summary data (headcount and wage bill).}
}
\description{
Overview Server Module
}
