% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_overview.R
\name{overview_server}
\alias{overview_server}
\title{Overview Server Module}
\usage{
overview_server(id, meso_data)
}
\arguments{
\item{id}{Module id.}

\item{workforce_data}{Data frame with workforce/personnel data (headcount).}

\item{wagebill_data}{Data frame with contract/salary data (wage bill).}

\item{cache}{List of cached data frames for improved performance.}
}
\description{
Overview Server Module
}
