% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_overview.R
\name{overview_server}
\alias{overview_server}
\title{Overview Server Module}
\usage{
overview_server(id, cache)
}
\arguments{
\item{id}{Module id.}

\item{cache}{List of cached data frames for improved performance.}
}
\description{
Overview Server Module
}
