% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_plots.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot Density as Percentage Share}
\usage{
plot_histogram(.data, plot_type = "histogram", group_col = NULL)
}
\arguments{
\item{.data}{A data frame.}

\item{plot_type}{A character string indicating the type of plot: "histogram" or "cumulative".}

\item{group_col}{The column name to group by.}
}
\value{
A plotly object.
}
\description{
Plot Density as Percentage Share
}
