% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_plots.R
\name{plot_transfer_heatmap}
\alias{plot_transfer_heatmap}
\title{Plot Transfer Heatmap}
\usage{
plot_transfer_heatmap(.data)
}
\arguments{
\item{data}{A data frame.}

\item{group}{Character string. Grouping variable.}
}
\value{
A ggplot2 object representing a heatmap of transfer values between groups
}
\description{
Plot Transfer Heatmap
}
