% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_plots.R
\name{plot_transition_network}
\alias{plot_transition_network}
\title{Plot Transition Network}
\usage{
plot_transition_network(.data)
}
\arguments{
\item{.data}{A data frame containing the transition data with columns \code{from}, \code{to}, and \code{weight}.}
}
\value{
A ggplot2 object representing a transition network.
}
\description{
Plot Transition Network
}
