% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govhr_port.R
\name{project_retirement}
\alias{project_retirement}
\title{Project Retirement Dates}
\usage{
project_retirement(
  workforce_data,
  threshold_age = 60,
  birth_col = "birth_date",
  group_cols = NULL,
  simplify_retirement_date = TRUE
)
}
\arguments{
\item{workforce_data}{A data frame containing personnel data with birth dates and reference dates.}

\item{threshold_age}{The age at which personnel are considered eligible for retirement (default is 60).}

\item{birth_col}{The name of the column representing birth dates (default is "birth_date").}

\item{group_cols}{A character vector of column names to group the data by when counting eligible retirees (default is NULL, meaning no grouping).}

\item{simplify_retirement_date}{A logical value indicating whether to simplify the retirement date to the end of the year (default is TRUE).}
}
\value{
A data frame with projected retirement dates and counts of staff eligible for retirement at each reference date.
}
\description{
Project Retirement Dates
}
\details{
The function takes a data frame containing personnel data with birth dates and reference dates, calculates the projected retirement date for each staff member based on the specified threshold age, and counts the number of staff eligible for retirement at each reference date.
}
