% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_utils.R
\name{render_movement_box}
\alias{render_movement_box}
\title{Render Movement Value Box}
\usage{
render_movement_box(.data, type_movement)
}
\arguments{
\item{.data}{A data frame containing personnel data.}

\item{type_movement}{A character string indicating the type of movement: "hire", "fire", or "turnover".}
}
\value{
A Shiny UI output for the movement value box.
}
\description{
Render Movement Value Box
}
