% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_utils.R
\name{render_movement_profile}
\alias{render_movement_profile}
\title{Render a movement profile table}
\usage{
render_movement_profile(data, movement_type)
}
\arguments{
\item{data}{A data frame containing personnel data with columns for ref_date, personnel_id, gender, educat7, employment_status, and birth_date.}

\item{movement_type}{A character string indicating the type of movement to profile (e.g
"hire" or "separation").}
}
\value{
A gt table summarizing the demographic characteristics of the specified movement type compared to the general population.
}
\description{
Render a movement profile table
}
