% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_utils.R
\name{render_wagebill_box}
\alias{render_wagebill_box}
\title{Function to render a wagebill summary value box.}
\usage{
render_wagebill_box(wagebill_data, type_measure)
}
\arguments{
\item{wagebill_data}{A data frame containing wagebill data.}

\item{type_measure}{A character string specifying which measure to display. Must be one of "total_wagebill" or "total_pension_liabilities".}
}
\value{
A Shiny UI output rendering a value box summarizing the wage bill or pension liabilities total.
}
\description{
Function to render a wagebill summary value box.
}
