% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_app.R
\name{run_govhrapp_qcheck}
\alias{run_govhrapp_qcheck}
\title{Run the govhr Shiny Dashboard Application}
\usage{
run_govhrapp_qcheck(
  est_data,
  personnel_data,
  contract_data,
  personnel_validation,
  contract_validation,
  ...
)
}
\arguments{
\item{est_data}{Data frame with establishment attributes.}

\item{personnel_data}{Data frame with personnel attributes.}

\item{contract_data}{Data frame with contract attributes.}

\item{personnel_validation}{List. The personnel validation results, including reports and violations.}

\item{contract_validation}{List. The contract validation results, including reports and violations.}

\item{...}{Additional arguments passed to \code{\link[shiny]{shinyApp}}.}
}
\value{
A Shiny app object.
}
\description{
Launches an interactive Shiny dashboard for govhr data quality check.
}
\examples{
\dontrun{
run_govhrapp_qcheck(est_data, personnel_data, contract_data)
}

}
