% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcheck_app.R
\name{run_qcheckapp}
\alias{run_qcheckapp}
\title{Run the Quality Assessment Shiny Application}
\usage{
run_qcheckapp(dynamicqc_obj, ...)
}
\arguments{
\item{dynamicqc_obj}{A dynamicqc object created by \code{\link{compute_dynamicqc}}.}

\item{...}{Additional arguments passed to \code{\link[shiny]{shinyApp}}.}
}
\value{
A Shiny app object.
}
\description{
Launches an interactive Shiny dashboard for govhr data visualization and analysis.
}
\examples{
\dontrun{
qc_data <- compute_dynamicqc(
  contract_dt = govhr::bra_hrmis_contract,
  personnel_dt = govhr::bra_hrmis_personnel,
  est_dt = govhr::bra_hrmis_est
)
run_qcheckapp(dynamicqc_obj = qc_data)
}

}
