% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{subgroup_filter_ui}
\alias{subgroup_filter_ui}
\title{UI for the subgroup filter input.}
\usage{
subgroup_filter_ui(id, .data)
}
\arguments{
\item{id}{Character string. The module namespace ID.}

\item{.data}{Data frame. The data to be used for determining the available subgroup choices.}
}
\value{
A Shiny UI element for selecting a subgroup filter, conditionally displayed based on the selected group.
}
\description{
UI for the subgroup filter input.
}
