% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_helpers.R
\name{update_group_filter_controls}
\alias{update_group_filter_controls}
\title{Update filter controls based on selected group variable}
\usage{
update_group_filter_controls(.data, input, session)
}
\arguments{
\item{.data}{A data frame containing the data to be filtered.}

\item{input}{A Shiny input object containing the selected group variable.}

\item{session}{A Shiny session object.}
}
\value{
None. This function updates the filter controls in the Shiny UI.
}
\description{
Update filter controls based on selected group variable
}
