% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_equity.R
\name{wagebill_equity_server}
\alias{wagebill_equity_server}
\title{Function to create the server logic for the wagebill equity module.}
\usage{
wagebill_equity_server(id, .data, cache)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}

\item{cache}{A list of cached data frames for improved performance.}
}
\value{
A Shiny module server function for the wagebill equity module.
}
\description{
Function to create the server logic for the wagebill equity module.
}
