% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_equity.R
\name{wagebill_equity_ui}
\alias{wagebill_equity_ui}
\title{Function to create the UI for the wagebill equity module.}
\usage{
wagebill_equity_ui(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}
}
\value{
A Shiny module UI function for the wagebill equity module.
}
\description{
Function to create the UI for the wagebill equity module.
}
