% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_movement.R
\name{wagebill_movement_server}
\alias{wagebill_movement_server}
\title{Function to create the server logic for the wagebill movement module.}
\usage{
wagebill_movement_server(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}
}
\value{
A Shiny module server function for the wagebill movement module.
}
\description{
Function to create the server logic for the wagebill movement module.
}
