% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_helpers.R
\name{wagebill_overview_server}
\alias{wagebill_overview_server}
\title{Function to create the server logic for the wagebill overview module.}
\usage{
wagebill_overview_server(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}
}
\value{
A Shiny module server function for the wagebill overview module.
}
\description{
Function to create the server logic for the wagebill overview module.
}
