% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_helpers.R
\name{wagebill_overview_ui}
\alias{wagebill_overview_ui}
\title{Function to identify available grouping choices based on the columns present in the data.}
\usage{
wagebill_overview_ui(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}
}
\value{
A list of UI elements for filtering and grouping wagebill data.
}
\description{
Function to identify available grouping choices based on the columns present in the data.
}
