% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_helpers.R
\name{wagebill_overview_ui}
\alias{wagebill_overview_ui}
\title{Function to create the UI for the wagebill overview module.}
\usage{
wagebill_overview_ui(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}
}
\value{
A Shiny module UI function for the wagebill overview module.
}
\description{
Function to create the UI for the wagebill overview module.
}
