% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_retirement.R
\name{wagebill_retirement_server}
\alias{wagebill_retirement_server}
\title{Server for the wagebill retirement module.}
\usage{
wagebill_retirement_server(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}
}
\value{
A Shiny module server function for the wagebill retirement module.
}
\description{
Server for the wagebill retirement module.
}
