% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill_helpers.R
\name{wagebill_retirement_ui}
\alias{wagebill_retirement_ui}
\title{Wagebill retirement UI module.}
\usage{
wagebill_retirement_ui(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}
}
\value{
A Shiny module UI function for the wagebill retirement module.
}
\description{
Wagebill retirement UI module.
}
