% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill.R
\name{wagebill_server}
\alias{wagebill_server}
\title{Wage Bill Server Module}
\usage{
wagebill_server(id, wagebill_data, cache)
}
\arguments{
\item{id}{Module id.}

\item{wagebill_data}{Data frame with wage bill data.}

\item{cache}{A list containing pre-computed trend summaries for workforce and wagebill data.}
}
\description{
Server logic for wage bill analytics server
}
