% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_wagebill.R
\name{wagebill_ui}
\alias{wagebill_ui}
\title{Wage Bill UI Module}
\usage{
wagebill_ui(id, wagebill_data)
}
\arguments{
\item{id}{Module id.}

\item{wagebill_data}{Data frame with wage bill data.}
}
\description{
UI for wage bill analytics, including overview, controls, and plots.
}
