% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_movement.R
\name{workforce_movement_server}
\alias{workforce_movement_server}
\title{Workforce Movement Server}
\usage{
workforce_movement_server(id, .data, cache)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing personnel data.}

\item{cache}{A list containing cached data for the module.}
}
\value{
A Shiny server function for the workforce movement module.
}
\description{
Workforce Movement Server
}
