% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_helpers.R
\name{workforce_movement_server}
\alias{workforce_movement_server}
\title{Workforce Movement Server}
\usage{
workforce_movement_server(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing personnel data.}
}
\value{
A Shiny server function for the workforce movement module.
}
\description{
Workforce Movement Server
}
