% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_helpers.R
\name{workforce_movement_ui}
\alias{workforce_movement_ui}
\title{Workforce Movement UI}
\usage{
workforce_movement_ui(id, .data, type_movement = c("hire", "fire"))
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing personnel data.}

\item{type_movement}{A character string specifying the type of movement: "hire" or "fire".}
}
\value{
A Shiny UI function for the workforce movement module.
}
\description{
Workforce Movement UI
}
