% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_movement.R
\name{workforce_movement_ui}
\alias{workforce_movement_ui}
\title{Workforce Movement UI}
\usage{
workforce_movement_ui(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing personnel data.}
}
\value{
A Shiny UI function for the workforce movement module.
}
\description{
Workforce Movement UI
}
