% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_helpers.R
\name{workforce_overview_server}
\alias{workforce_overview_server}
\title{Workforce Overview Server}
\usage{
workforce_overview_server(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing workforce data.}
}
\value{
A Shiny server function for the workforce overview module.
}
\description{
Workforce Overview Server
}
