% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_overview.R
\name{workforce_overview_ui}
\alias{workforce_overview_ui}
\title{Workforce Overview UI}
\usage{
workforce_overview_ui(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing workforce data.}
}
\value{
A Shiny UI function for the workforce overview module.
}
\description{
Workforce Overview UI
}
