% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_retirement.R
\name{workforce_retirement_server}
\alias{workforce_retirement_server}
\title{Workforce Retirement Server}
\usage{
workforce_retirement_server(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing personnel data.}
}
\value{
A Shiny server function for the workforce retirement module.
}
\description{
Workforce Retirement Server
}
