% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_helpers.R
\name{workforce_retirement_ui}
\alias{workforce_retirement_ui}
\title{Workforce Retirement UI}
\usage{
workforce_retirement_ui(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing personnel data.}
}
\value{
A Shiny UI function for the workforce retirement module.
}
\description{
Workforce Retirement UI
}
