% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce.R
\name{workforce_server}
\alias{workforce_server}
\title{Workforce Analytics Server Module}
\usage{
workforce_server(id, workforce_data, wagebill_data, cache)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{workforce_data}{A data frame containing personnel data.}

\item{wagebill_data}{A data frame containing wage bill data.}

\item{cache}{A list containing pre-computed trend summaries for workforce data.}
}
\value{
A Shiny module server function for workforce analytics.
}
\description{
Workforce Analytics Server Module
}
