% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce.R
\name{workforce_server}
\alias{workforce_server}
\title{Workforce Server Module}
\usage{
workforce_server(id, workforce_data)
}
\arguments{
\item{id}{Module id.}

\item{workforce_data}{Data frame with workforce data.}
}
\description{
Server logic for workforce analytics UI.
}
