% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_transition.R
\name{workforce_transition_server}
\alias{workforce_transition_server}
\title{Workforce Transition Server Module}
\usage{
workforce_transition_server(id, .data, cache)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}

\item{cache}{A list containing pre-computed data for caching.}
}
\value{
A Shiny server module for workforce transition analytics.
}
\description{
Workforce Transition Server Module
}
