% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_utils.R
\name{workforce_transition_ui}
\alias{workforce_transition_ui}
\title{Workforce Transition UI}
\usage{
workforce_transition_ui(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing workforce data.}
}
\value{
A Shiny UI function for the workforce transition module.
}
\description{
Workforce Transition UI
}
