% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce_transition.R
\name{workforce_transition_ui}
\alias{workforce_transition_ui}
\title{Workforce Transition UI Module}
\usage{
workforce_transition_ui(id, .data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{.data}{A data frame containing wagebill data.}
}
\value{
A Shiny module UI function for workforce transition analytics.
}
\description{
Workforce Transition UI Module
}
