% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce.R
\name{workforce_ui}
\alias{workforce_ui}
\title{Workforce Analytics Module}
\usage{
workforce_ui(id, workforce_data, wagebill_data)
}
\arguments{
\item{id}{A character string specifying the module ID.}

\item{workforce_data}{A data frame containing personnel data.}

\item{wagebill_data}{A data frame containing wage bill data.}
}
\value{
A Shiny app object for workforce analytics.
}
\description{
Workforce Analytics Module
}
