% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics_workforce.R
\name{workforce_ui}
\alias{workforce_ui}
\title{Workforce UI Module}
\usage{
workforce_ui(id, workforce_data)
}
\arguments{
\item{id}{Module id.}

\item{workforce_data}{Data frame with workforce data.}
}
\description{
UI for workforce analytics, including overview, controls, and plots.
}
